requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Fetchers
requests>=2.31.0
beautifulsoup4>=4.12.0

# Dashboard
streamlit>=1.40.0
